requirements.txt
# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Data Generation
faker==22.6.0
